streaming_alert/orchestrator/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-firestore==2.17.0
google-cloud-bigquery==3.25.0
